+++ nanocv/layers/layer_linear.cpp
@@ -66,9 +66,9 @@ namespace ncv
 
         const tensor_t& linear_layer_t::output(const tensor_t& input)
         {
-                assert(idims() == input.dims());
-                assert(irows() == input.rows());
-                assert(icols() == input.cols());
+                assert(idims() == static_cast<size_t>(input.dims()));
+                assert(irows() == static_cast<size_t>(input.rows()));
+                assert(icols() == static_cast<size_t>(input.cols()));
 
                 m_idata = input;
 
@@ -79,9 +79,9 @@ namespace ncv
 
         const tensor_t& linear_layer_t::ginput(const tensor_t& output)
         {
-                assert(output.dims() == odims());
-                assert(output.rows() == orows());
-                assert(output.cols() == ocols());
+                assert(static_cast<size_t>(output.dims()) == odims());
+                assert(static_cast<size_t>(output.rows()) == orows());
+                assert(static_cast<size_t>(output.cols()) == ocols());
 
                 m_odata = output;
 
@@ -92,9 +92,9 @@ namespace ncv
 
         void linear_layer_t::gparam(const tensor_t& output, scalar_t* gradient)
         {
-                assert(output.dims() == odims());
-                assert(output.rows() == orows());
-                assert(output.cols() == ocols());
+                assert(static_cast<size_t>(output.dims()) == odims());
+                assert(static_cast<size_t>(output.rows()) == orows());
+                assert(static_cast<size_t>(output.cols()) == ocols());
 
                 m_odata = output;
