+++ nanocv/layers/layer_pool.cpp
@@ -33,9 +33,9 @@ namespace ncv
 
         const tensor_t& pool_layer_t::output(const tensor_t& input)
         {
-                assert(idims() == input.dims());
-                assert(irows() <= input.rows());
-                assert(icols() <= input.cols());
+                assert(idims() == static_cast<size_t>(input.dims()));
+                assert(irows() <= static_cast<size_t>(input.rows()));
+                assert(icols() <= static_cast<size_t>(input.cols()));
 
                 m_idata = input;
 
@@ -54,9 +54,9 @@ namespace ncv
 
         const tensor_t& pool_layer_t::ginput(const tensor_t& output)
         {
-                assert(odims() == output.dims());
-                assert(orows() == output.rows());
-                assert(ocols() == output.cols());
+                assert(odims() == static_cast<size_t>(output.dims()));
+                assert(orows() == static_cast<size_t>(output.rows()));
+                assert(ocols() == static_cast<size_t>(output.cols()));
 
                 m_odata = output;
 
@@ -78,9 +78,9 @@ namespace ncv
                 NANOCV_UNUSED1(gradient);
                 NANOCV_UNUSED1_RELEASE(output);
 
-                assert(odims() == output.dims());
-                assert(orows() == output.rows());
-                assert(ocols() == output.cols());
+                assert(odims() == static_cast<size_t>(output.dims()));
+                assert(orows() == static_cast<size_t>(output.rows()));
+                assert(ocols() == static_cast<size_t>(output.cols()));
         }
 }
